+++ src/advanced/features/product-management/index.ts
@@ -1,0 +1,2 @@
+export { useProductService } from './model/useProductService';
+export { useProductForm } from './ui/useProductForm';
